+++ src/main/java/ru/evotor/framework/inventory/product/extension/AlcoholProduct.kt
@@ -6,5 +6,5 @@ import java.math.BigDecimal
 interface AlcoholProduct : ProductExtension {
     val fsrarProductKindCode: Long
     val tareVolume: AmountOfLiters
-    val alcoholPercentage: BigDecimal?
+    val alcoholPercentage: Float?
 }
